bff/requirements.txt
python-decouple==3.8
spacy==3.6.1
aiofiles~=22.1.0
beautifulsoup4~=4.12.3
fastapi==0.103.2
geopy==2.4.0
motor==3.6.0
opencage==2.3.0
pandas==2.2.3
pydantic==2.3.0
pymongo~=4.9.2
requests==2.31.0
uvicorn==0.23.2
pika~=1.3.2
aio_pika
kombu
pytest~=8.3.4
python-socketio